created custommongoose helper that function to check if objectId is valid

+++ backend/controllers/bucketlistController.js
@@ -1,6 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const Bucketlist = require('../models/bucketlistModel');
-
+const { objectIdCheck } = require('../helpers/custommongooseHelper');
 // @desc    Get >bucketlist
 // @route   GET /api/bucketList
 // @access  Private
@@ -27,13 +27,19 @@ module.exports.post = asyncHandler(async (req, res) => {
   res.status(201).json(bucketlist);
 });
 
+// update & delete to do : handle if req.params.id inserted are not a valid object id (error thrown by mongoose :Cast to ObjectId failed for value \"req.params.id\" (type string) at path \"_id\" for model \"bucketlists\")
+
 // @desc    Update >bucketlist
 // @route   PUT /api/bucketList/:id
 // @access  Private
 module.exports.put = asyncHandler(async (req, res) => {
+  if (!objectIdCheck(req.params.id)) {
+    res.status(404);
+    throw new Error('Id format invalid');
+  }
   // check if id exist and update data in db, third argument is option to create if doesnt exist
   const updatedBucketlist = await Bucketlist.findByIdAndUpdate(
-    req.params.id,
+    { _id: req.params.id },
     req.body,
     { new: true }
   );
@@ -49,6 +55,10 @@ module.exports.put = asyncHandler(async (req, res) => {
 // @route   DELETE /api/bucketList/:id
 // @access  Private
 module.exports.delete = asyncHandler(async (req, res) => {
+  if (!objectIdCheck(req.params.id)) {
+    res.status(404);
+    throw new Error('Id format invalid');
+  }
   // deleting data on db
   const deletedBucketlist = await Bucketlist.findByIdAndDelete(req.params.id);
   if (!deletedBucketlist) {
